Removed setup_requires = ['persistent'] since the persistent module is not actually needed at setup time- it is only needed in install_requires.
When in setup_requires, this breaks configurations that use local repositories
like Artifactory. There are no clean workarounds that I have been able to find.

I tested the installer by just removing setup_requires and it works fine. I don't
understand why it would need to be there in the first place. Maybe it was a mistake?