Avoid spurious "failures" on Appveyor during releases
PowerShell interprets any stderr output from a child process as an
error.  This means warnings from pip about using
not-the-very-latest-version-how-dare-you and warnings from twine about
uploading the *.tar.gz which was probably already uploaded already and
might not be a bitwise identical copy because what even is
reproducibility cause the Appveyor builds to fail, even if the actual
binary wheels we want have been built and uploaded successfully.

*pauses to breathe*