Register BTree/Bucket as MutableMapping and Set/TreeSet as MutableSet.
Add tests to verify they implement the required methods. Turns out they didn't, so implement them.
This means popitem for the mappings, and pop/discard/isdisjoint plus a swath of operators for the sets.

Do this in both C and Python. The implementations are probably on the naive side, in that they work with generic iterators and don't try to do any complex structural merging,
but that makes them easier to understand and hopefully be correct. Performance shouldn't be too bad, nothing worse than O(N*logM) (roughly). One method requires extra storage space of O(N), and __xor__ requires storage of O(N*M).

Fixes #121