Expose max_internal_size and max_leaf_size as rw on the C classes
Just like on the Python clasess. Fixes #166

This takes a metaclass at the C level, but it's a very simple one.

In addition to consistency, and letting the sizes be customized for an
entire application, this has some nice properties:

- We can optimize some zope.interface object storage
- We delete the DEFAULT_MAX_*_SIZE macros. Now there's only one source
  of truth for those, whether in C or Python: _datatypes.py

In addition, I was able to make a small optimization for
__slotnames__. Previously it was computed (to be empty) and then
discarded over and over (every time you pickled or deactivated an
object); now it is properly cached. This won't affect any subclasses.